Fix issues 459, 460, 461
configure_arguments is now populated in the database submitter

MPIVersion marks the name and version of MPI as "None" and "Unknown" if mpicc is not found

get_mpi_version is auto-generated so that the user running MTT has access to the file
