+++ pylib/Utilities/MPIVersion.py
@@ -30,6 +30,11 @@ class MPIVersion(BaseMTTUtility):
     def execute(self, log, testDef):
 
         version_str = self.get_version_string(testDef)
+
+        if version_str is None:
+            log['name'] = 'None'
+            log['version'] = 'Unknown'
+            return
         
         name = None
         version = None
@@ -67,17 +72,44 @@ class MPIVersion(BaseMTTUtility):
         return
 
     def get_version_string(self, testDef):
-        cwd = os.getcwd()
-        os.chdir(os.path.dirname(os.path.realpath(sys.modules[self.__module__].__file__)))
+
+        if not os.path.isdir(os.path.join(os.getcwd(), "mttscratch")):
+            os.mkdir("mttscratch")
+        os.chdir("mttscratch")
+        fh = open("mpi_get_version.c", "w")
+        fh.write("""
+/* This program is automatically generated by MPIVersion.py
+ * of MPI Testing Tool (MTT). Any changes you make here may
+ * get lost!
+ * Copyrights and licenses of this file are the same as for the MTT.
+ */
+#include <mpi.h>
+#include <stdio.h>
+int main(int argc, char **argv) {
+    MPI_Init(NULL, NULL);
+    char version[1000];
+    int resultlen;
+    MPI_Get_library_version(version, &resultlen);
+    printf("%s\\n", version);
+    return 0;
+}""")
+        fh.close()
         status, _, _ = testDef.execmd.execute('mpicc -o mpi_get_version mpi_get_version.c'.split(), testDef)
         if 0 != status:
-            os.chdir(cwd)
+            if os.path.exists("mpi_get_version"): os.remove("mpi_get_version")
+            if os.path.exists("mpi_get_version.c"): os.remove("mpi_get_version.c")
+            os.chdir("..")
             return None
+
         status, stdout, _ = testDef.execmd.execute('./mpi_get_version'.split(), testDef)
         if 0 != status:
-            os.chdir(cwd)
+            if os.path.exists("mpi_get_version"): os.remove("mpi_get_version")
+            if os.path.exists("mpi_get_version.c"): os.remove("mpi_get_version.c")
+            os.chdir("..")
             return None
         
-        os.chdir(cwd)
-        return stdout
+        if os.path.exists("mpi_get_version"): os.remove("mpi_get_version")
+        if os.path.exists("mpi_get_version.c"): os.remove("mpi_get_version.c")
+        os.chdir("..")
+        return str(stdout)
